Use --app-namespace rather than -n for dedicated option in state-namespace.md
When using a dedicated namespace for state storage, the docs currently suggest specifying this namespace via the `-n` flag. However this can lead to undesirable behaviors, since the app is typically being deployed to a different namespace. In my case, I had a resource where I had neglected to specify the namespace, and it would have ended up getting deployed to the state-storage namespace rather than the same namespace as the rest of the app.

This was discussed in https://github.com/carvel-dev/kapp/issues/815 and fixed in https://github.com/carvel-dev/kapp/pull/814 with the introduction of the `--app-namespace` flag, which controls which is used for state storage independently of the app namespace. However the docs did not reflect this change.

This commit adjusts the documentation to recommend using `--app-namespace` when storing configs in a separate namespace.